refactor(platform): improve token distribution tests
